ingest/persons_data/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.20.1
google-cloud-storage==2.16.0
requests==2.31.0